Covering LogicTypeSelector
