+++ kie-wb-common-react/boxed-expression-editor/src/__tests__/components/ExpressionContainer/ExpressionContainer.test.tsx
@@ -15,14 +15,10 @@
  */
 
 import { ExpressionContainer } from "../../../components/ExpressionContainer";
-import { fireEvent, render } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import * as React from "react";
 import { usingTestingBoxedExpressionI18nContext } from "../test-utils";
 import { DataType, LogicType } from "../../../api";
-import { act } from "react-dom/test-utils";
-
-jest.useFakeTimers();
-const flushPromises = () => new Promise((resolve) => process.nextTick(resolve));
 
 describe("ExpressionContainer tests", () => {
   test("should render ExpressionContainer component", () => {
@@ -65,74 +61,4 @@ describe("ExpressionContainer tests", () => {
     expect(container.querySelector(".expression-type")).toBeTruthy();
     expect(container.querySelector(".expression-type")!.innerHTML).toBe("(&lt;Undefined&gt;)");
   });
-
-  describe("Expression Actions dropdown", () => {
-    test("should have the clear action disabled on startup", async () => {
-      const expression = { name: "Test", dataType: DataType.Undefined };
-
-      const { container } = render(
-        usingTestingBoxedExpressionI18nContext(<ExpressionContainer selectedExpression={expression} />).wrapper
-      );
-
-      await triggerContextMenu(container as HTMLElement, ".logic-type-selector");
-
-      expect(container.querySelector(".context-menu-container button.pf-m-disabled")).toBeTruthy();
-      expect(container.querySelector(".context-menu-container button.pf-m-disabled")!.innerHTML).toBe("Clear");
-    });
-
-    test("should have the clear action enabled, when logic type is selected", async () => {
-      const expression = { name: "Test", logicType: LogicType.LiteralExpression, dataType: DataType.Undefined };
-
-      const { container } = render(
-        usingTestingBoxedExpressionI18nContext(<ExpressionContainer selectedExpression={expression} />).wrapper
-      );
-
-      await triggerContextMenu(container as HTMLElement, ".logic-type-selector");
-
-      expect(container.querySelector(".context-menu-container button.pf-m-disabled")).toBeFalsy();
-      expect(container.querySelector(".context-menu-container button")).toBeTruthy();
-      expect(container.querySelector(".context-menu-container button")!.innerHTML).toBe("Clear");
-    });
-  });
-
-  describe("Logic type selection", () => {
-    test("should show the pre-selection, when logic type prop is passed", () => {
-      const expression = { name: "Test", logicType: LogicType.Invocation, dataType: DataType.Undefined };
-
-      const { container } = render(
-        usingTestingBoxedExpressionI18nContext(<ExpressionContainer selectedExpression={expression} />).wrapper
-      );
-
-      expect(container.querySelector(".expression-container-box")).toBeTruthy();
-      expect(container.querySelector(".expression-container-box")!.innerHTML).toContain(expression.logicType);
-    });
-
-    test("should reset the selection, when logic type is selected and clear button gets clicked", async () => {
-      const expression = { name: "Test", logicType: LogicType.LiteralExpression, dataType: DataType.Undefined };
-
-      const { container } = render(
-        usingTestingBoxedExpressionI18nContext(<ExpressionContainer selectedExpression={expression} />).wrapper
-      );
-
-      await triggerContextMenu(container as HTMLElement, ".logic-type-selector");
-
-      act(() => {
-        const clearButtonElement = container.querySelector(".context-menu-container button")!;
-        const clearButton = clearButtonElement as HTMLButtonElement;
-        clearButton.click();
-      });
-
-      expect(container.querySelector(".expression-container-box")).toBeTruthy();
-      expect(container.querySelector(".expression-container-box")!.innerHTML).not.toBe(expression.logicType);
-    });
-  });
 });
-
-const triggerContextMenu = async (container: HTMLElement, selector: string) => {
-  await act(async () => {
-    const element = container.querySelector(selector)!;
-    fireEvent.contextMenu(element);
-    await flushPromises();
-    jest.runAllTimers();
-  });
-};
